py/requirements.txt
bayesian-optimization==1.4.3
colorama==0.4.6
grpcio==1.60.0
grpcio-tools==1.60.0
joblib==1.3.2
numpy==1.26.3
protobuf==4.25.2
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
threadpoolctl==3.2.0
